+++ src/modules/index.ts
@@ -4,3 +4,4 @@ export * from "./main";
 export * from "./unit";
 export * from "./aminoAcid";
 export * from "./fattyAcid";
+export * from "./nutrient";
