+++ src/modules/nutrient/index.ts
@@ -1,0 +1,1 @@
+export * from "./schema";
